+++ internal/parsergen/backend/rust/doc.go
@@ -1,0 +1,2 @@
+// Package rust provides a parser generator backend which takes a parse table as input and outputs Rust source code.
+package rust